+++ src/map/skills/other/netsupport.cpp
@@ -14,7 +14,7 @@ void SkillNetSupport::castendNoDamageId(block_list* src, block_list* target, uin
 	status_data* tstatus = status_get_status_data(*target);
 
 	if (flag & 1) {
-		int32 heal_amount = tstatus->max_sp * 3 / 100;
+		int32 heal_amount = static_cast<int64>(tstatus->max_sp) * 3 / 100;
 		clif_skill_nodamage(nullptr, *target, MG_SRECOVERY, heal_amount);
 		status_heal(target, 0, heal_amount, 0);
 	} else {
